Removed generics from SOM abstract classes and functions.
On begining to implement the SOM functions for CDK IAtomContainers
I ran into a problem with the type system. I needed the types of the
Fields and training data to be different (Field<List<T>> and T) but
attempting to ensure type safety was ugly (SOM<T1,T2> having to be
propagated, and managed) and added mental overhead.

Now, we'll have to relie on casting and runtime checking.